Bring solarized themes into parity with official
Because the solarized colors from these themes are not organized the
same way as the xresources from the solarized repository, using these
xresources would break themes for programs written against the
xresources in the solarized repository. For example, the solarized theme
for irssi becomes nearly unreadable when using the base16-solarized
xresources.

This commit re-organizes the solarized xresources to match the
xresources from the official solarized repository. As such, this commit
will break the base16-solarized vim theme, and any other themes relying
on base16's version of xresource organization.

With that said, this commit may not be inline with the goals of the
base16 project; however, I wanted to at least start a discussion. In my
testing, I've found the organization from the solarized project to
provide better readability overall, for both themed programs and general
terminal usage, even when applied to other color schemes. For example,
when I copied the assignments from this updated solarized-dark xresource
file to the default-dark xresource file, general terminal readability
improved.

Perhaps this is just a coincidence of my particular environment, but if
not, it might be worth changing.